Add height hack option
The display we are going to be using is 1980x360,
but it reports as a 1980x1020 display and I can't
figure out how to configure X to limit the screen
to this size. This hack will still render to the "full"
screen, but the important text will show up where
we want it.